Fixes CB-101 can't access media in documents://subDir
Needed to rework urlForResource in sound.m to test for http and documents prefixes
before looking in web folder for media resource. Also needed to modify how file path
is created in case of documents:// resource.  Added check for existence of all file type
media and error if file does not exist.